refactor: merge pointer.py into binding.py
Pointer utilities (is_pointer, parse_pointer, scan_for_pointers,
PointerInfo) are conceptually part of the binding infrastructure.
Merge into binding.py and remove the standalone pointer module.

583 tests pass, ruff clean, coverage 91%.
